Low: scheduler: require pre-/post-start notifications correctly
4974f08 in 2.0.1 introduced a regression in the notification code by freeing
and nulling a variable (rsc_list) before it was checked against NULL later.

The check was never even needed to begin with, because rsc_list could never be
NULL at that point (before the regression).

The effect of the regression is that pre-/post-start notify actions would not
automatically become required actions when there were instances starting.
However they seem to become required elsewhere, so it probably was not
significant.