ipc_client: Use libqb to get auth data for ipcc connections
When using libabq in SOCKET mode it uses DGRAM connections
that don't support SO_PEERCRED, so we can't use that to authenticate
the daemons. However, libqb does keep this information so we can
use the (newly created) API call to get it.

This does not work if corosync is using SOCKET connections (set
in corosync.conf) as it does not expose the libqb ipcc object
we need to pass into libqb.