Fix: tools: Simplify cib connection tear down in crm_mon.
Just use a single function for this, and move the error/info printing
out into the callers.  In preparation for printing any stonith history
error messages as part of crm_mon's standard output, I want to remove
any other printing from the loop in mon_refresh_display.  This is part
of it.