Feature: tools: Use formatted output for CTS printing in crm_resource.
Note that this option only exists for internal use, passing information
to cts.  It's not exposed in the command line help.  I've implemented it
just using out->info so it will only print for the text output format.
I don't want XML output that someone might start to rely on, despite not
being in the schema.