Feature: tools: Use formatted output for stacks and constraints.
This also changes the format for text output.  The new output uses the
underlying text list code for handling indentation and nesting, while
the old code passed around an indentation prefix.  However, this does
mean we end up with a little of the text list fanciness as well.
Hopefully that will not be a problem.