Fix: Guard config.h against multiple inclusion.
With the gnulib update, a bunch more _GL_ATTRIBUTE_* macros are now
being defined in config.h.  That header file does not have any guard in
it to protect against multiple inclusion, so it's up to us to do so.
There's at least one place we were doing this - crm_internal.h includes
it directly as well as includes options_internal.h which also includes
it.