Test: cts-fencing: optimize fencing regression testing via appropriate synchronization with fenced
Previously for each test case, it would sleep for 1s waiting for fenced
to be initialized and ready for accepting connections. It would
accumulate and become expensive given that there are more than 60 test
cases.

With this commit, rather than sleeping for 1s, it tracks the log and
starts each test case whenever fenced is shown being ready. So that it
will save around 60s in total for the whole testing with an average
system (reduced from around 2m40s to 1m40s for my case).

On the other hand, for a system with poor performance, this solution
also prevents test failures that are due to impatience (1s sleep might
not be sufficient).