Test: cts-exec: optimize exec regression testing via appropriate synchronization with execd
Previously for each test case, it would sleep for 1s waiting for execd
to be initialized and ready for accepting connections. It would
accumulate and become expensive given that there are more than 40 test
cases.

With this commit, rather than sleeping for 1s, it tracks the log and
starts each test case whenever execd is shown being ready. So that it
will save around 40s in total for the whole testing with an average
system (reduced from around 7m15s to 6m35s for my case).

On the other hand, for a system with poor performance, this solution
also prevents test failures that are due to impatience (1s sleep might
not be sufficient).