Fix: libcrmcommon: Speed up parse_op_key.
If we were successful in finding exactly two underscores, pass their
locations to a special version of try_basic_match.  We can copy out
the parts of the key without having to go hunting.

Having done this, the interval conversion happens in too many places so
break that out into its own function.