Refactor: cib: simplify cib_process_request()
The force_synchronous argument was always FALSE, so ditch it.

The from_peer variable was equivalent to cib_client==NULL, so just use that
(the abundance of variables in this function is confusing).

host was unnecessarily re-declared and re-grabbed in one block.

is_update was unnecessarily checked in an else block after is_update==FALSE
had already been excluded earlier.

Also, add more comments.