Fix: cib: only build reply when needed
If a requester sets cib_discard_reply, we would still mark the request as
needing a reply if it modified the CIB, in case we needed to broadcast a diff.
However, that is only necessary in legacy mode, so only do it in that case.

This avoids some log error messages since d71fb9ac and f87436c about an
"Unexpected NULL reply to message" (at least when not in legacy mode,
which is only needed during rolling upgrades with some other node using
Pacemaker 1.1.11 or earlier).