Fix: scheduler: Remove pe_print_expanded_xml print option.
I had added this on the assumption that we would want to use the basic
resource formatted output messages for also printing out the XML config
of a resource.  However, we're going to go a different direction and
have the XML come from a separate output message.

So, this option can go away.  It was only ever in -rc1 so it's unlikely
anyone is using it.  I'm not simply reverting the patch that introduced
it because that patch also changed how all these pe_print_* options were
defined, and we'd like to keep that formatting.