Refactor: libpe_status: functionize creating operation digests better
Basically, this takes most of rsc_action_digest(), and puts it in a new,
exposed function pe__calculate_digests(), for future reuse. The exposed
function creates a new op_digest_cache_t object with calculated digests;
rsc_action_digests() takes that and puts it in a node's digest cache.

This additionally functionizes most of pe__calculate_digests(), with separate
functions for creating each of the three digests that go into a digest object:
the digest of all parameters, the digest of non-private parameters, and the
digest of reload parameters.

There are no changes in how the code works.