Test: scheduler: update regression test for promotion change
The promotion change in the previous commit affected an existing test for the
better. The summary and graph output did not actually change (other than action
IDs), just the process by which the graph was arrived, which can be summed up
in this trace output difference:

5260a5261
> (can_be_master)       trace: galera:1 cannot be promoted: guest galera-bundle-docker-1 not allocated
5262d5262
< (pcmk__set_instance_roles)    info: Promoting galera:1 (Master galera-bundle-1)
5266c5266
< (pcmk__set_instance_roles)    info: galera-bundle-master: Promoted 2 instances of a possible 3 to master
---
> (pcmk__set_instance_roles)    info: galera-bundle-master: Promoted 1 instances of a possible 3 to master

Previously, 2 instances were selected to be promoted, but only 1 of those
promotes ended up being runnable. Now, only the 1 runnable instance is selected
to begin with.

This shows up in the regression test differences as Slave monitors rather than
Master monitors being scheduled for the unrunnable instance (the monitors
themselves being unrunnable as well, there are no substantive graph changes).