Fix: tests: Use g_assert_true, g_assert_false, and g_assert_null.
These macros were defined in glib 2.38, which we can't use and it seems
like we won't be able to use it any time soon.  Luckily, it's easy
enough to define these in terms of other glib assertion macros in the
case we're on an older system.  That's basically what we were already
doing.  Now we'll only be doing it some of the time instead of all of
the time.