Low: scheduler: calculate secure digest properly for node attribute rules
6830621 corrected secure digest calculation in most cases, but did not work
properly when non-sensitive parameters depended on node attribute-based rules
(since it used rsc->parameters, which is not evaluated for node).

This fixes it by using the new resource parameters function (which is the
equivalent of what calculate_main_digest() already was doing, so it is now
exposed earlier for use by both digest functions).