Refactor: libcrmcluster: redo nodelist check
Previously, corosync_cmap_has_config() could check for any CMAP section name,
but it statically retained the result of its first call, so if it had actually
been called multiple times with different section names, the wrong result could
be returned.

Since in practice it was only called with "nodelist" as its argument, redo it
so it only checks for that. Also, improve its log messages, rename according to
current guidelines, and move to the library-private header.

This changes the behavior when an error occurs. Previously, the function would
return -1 on error, but since the lone caller only checked whether the return
value was false, an error would be treated as true. Now, an error is treated as
false (though there is no obvious ideal choice in that situation).