Low: libcrmcluster: properly check CPG payload size
For uncompressed CPG messages, pcmk_message_common_cs() called
check_message_sanity() with a "data" argument. However, "data" is always NULL
at that point. It is apparent that this was intended to be msg->data, so use
that directly in check_message_sanity() and drop the unnecessary argument.

Also remove a block of dead code in pcmk_message_common_cs() that couldn't be
reached because data was always NULL. There was no apparent use for this code.