Refactor: libs: Set XML properties when creating a node.
pcmk__output_create_xml_node can take a NULL-terminated list of
properties that should be set on the node at the same time it's created.
This saves a bunch of calls to xmlSetProp and gets rid of some variables
that are otherwise not needed.

Unfortunately, this can't be used everywhere I'd like but it's a good
start.