Refactor: libs: Add a function to set properties on an xmlNode.
This takes a va_list so it can be used by other future functions that
will take arguments directly.  It then just walks the list of pairs and
sets them as properties on an xmlNode, until a NULL is found.

Interally, it just uses crm_xml_add so it's useful across the whole
pacemaker code base.