Fix: libs, tools: Be consistent with GListPtr and xmlNodePtr in formatted output.
GListPtr should be removed from all the formatted output messages.  It
seems like a GLib type, but it's not.  So that's weird and we should
stop spreading uses of it around.

On the other hand, xmlNodePtr is a libxml type that we are already using
sporadically.  We should be more consistent in using it.