Fix: libs, tools: Add -list to certain formatted output message names.
Basically, if the message creates its own list, the name should end in
-list.  This is a hint to the caller that they do not need to add an
extra layer of lists.  I've further changed the names of these messages
to be a little shorter, too.

Note that some messages (the resource variants most obviously) create
their own list but don't have -list in the name.  Here, the list is more
for organizational purposes rather than indicating that it contains
items.