Fix: libs: Decrease the number of visible formatted output functions.
The only ones that need to be exported are the ones that crm_mon needs
to reference for its output.  All the others can be static and only
accessed indirectly through calling out->message.

And then remove the prefixes from these newly static functions.  This
could potentially be a little confusing if the text version is public
(and therefore has the, say, pe__ prefix) while the xml version is
hidden and is lacking that suffix.  But I think it'll be okay.