Refactor: scheduler: pull interval from digest overrides if appropriate
When calculating an operation digest, we previously used an operation key
provided by the caller, along with a table of parameter overrides. However, if
CRM_meta_interval is one of the overrides, that should be used in the operation
key instead of the caller-provided key.

Now, the caller passes in a pointer to the interval, and if it's overridden,
pe__calculate_digests() will use the overridden value and reset the caller's
interval to it.

As of this commit, no caller passes overrides, so it has no effect, but it will
be useful for an upcoming feature.