Build: libstonithd: add configure option for fence executable directory
New "--with-fence-bindir" configure option (defaulting to sbindir) allows
specifying where RHCS-style fence agents are located. (This is mainly useful
for CI testing and nonstandard installs.)

While we're at it, rename the corresponding C constant according to current
style guidelines, and move the non-configurable RH_STONITH_PREFIX constant from
the configure script to the only file that uses it.