Doc: Pacemaker Explained: Fix the node attributes expression table.
39d78f36b23d8910afdafd955df0170e0102b7e7 left the previous default in
the table, meaning we had a row with four columns, which ended up
shifting everything over by a column.  This is now fixed, the the
default column is so narrow that it's a little hard to read.

It doesn't matter too much, because this is all going away in favor of
new documentation soon.