Low: executor: properly initiate quicker fence device monitor
When merging a duplicate recurring monitor for a fencing device,
the code previously got the sense of a test wrong -- the monitor
needs to be initiated if it's *not* already pending. On top of that,
the code wrongly tried to schedule the new command rather than the
merged duplicate (the new command is destroyed immediately after this).