Fix: libpacemaker: Fix printing out location constraints.
In the text output, put the location constraints inside empty lists.
This will indent it under the previous line of text, making it more
apparent what the location information applies to.

In both text and XML output, remove the print out of the initial
resource.  There's plenty of context for what is happening already.  We
don't need an extra level.