Build: configure: default --enable-legacy-links to no
This has one change in behavior: a link will always be created for
pacemaker_remoted, regardless of the value of --enable-legacy-links. This
allows container images with Pacemaker >=2.0.0 to be used in bundles on
cluster nodes running Pacemaker 1 (>=1.1.17), which should be kept until we
wish to drop that backward compatibility (probably in a future 3.0.0 version).