Revert "Refactor: scheduler: functionize checking whether colocation applies"
This reverts commit 2278e01f8d951d939c172ac71e168a11199f84f7.

The original idea was that additional conditions (i.e. influence) could be
added to the new function, but influence should only affect whether the
dependent resource's location preferences are considered, not whether the
colocation constraint applies (i.e. the dependent should be stopped if it
can't run where the "with" resource is).