Low: tools: Adjust fencing shown indicator in crm_mon.
If any of the various fencing flags are set, but not all of them, no '*'
will be shown next to the fencing line in the interactive change screen.
This makes it seem like fencing should not be shown, and hitting 'm'
should toggle the fencing display on.  However, that's not the case and
hitting 'm' will actually toggle fencing off.  Hitting it again will
toggle it on and the '*' will appear.

This is confusing, so just display the '*' if any fencing flag is set.