Build: CIB: remove cibmon command
This was installed in libexec and so was not intended for end users.

It originated in the (pre-Pacemaker) heartbeat project, for which it would log
CIB diffs. Pacemaker logs CIB diffs to the detail log automatically, so that
is not useful. Otherwise it served primarily as an example of how to use the
CIB IPC API, which again is not useful.