Build: configure: require getopt() to be available
Previously, we checked for getopt()'s existence, which would create cache
variables indicating that, but we did nothing with the variables, and we
use getopt() unconditionally in the code.

Now, define a REQUIRE_FUNC() m4 macro for reuse, that errors out if a function
isn't available in the standard C library (actually the LIBS available at that
point). Use that to require getopt().