Test: scheduler: add regression test for remote connection shutdown
In particular, a remote node is shutting down, and its node history comes
before a cluster node's history in the status section, where that cluster node
has a probe that found the remote node's connection not running in its history.

The test ensures that the connection is scheduled to be stopped, even though
that later history will set the next role to unknown.