Low: pacemakerd: use high-resolution sleep functions if available
In configure, detect whether nanosleep() and usleep() are available.

In libcrmcommon, define an internal API for sub-second sleep that uses
nanosleep(), usleep(), or select() as available.

In pacemakerd, use the new function to sleep instead of poll().

This is not hugely significant but is portable and readable, and allows us to
use sub-second sleeps elsewhere in the future if desired.