Low: tools: Allow other output formats for crm_mon -d.
There's no reason we can't have text output for daemonize mode.  We just
need to specifically allow it (by not resetting the format to none) and
then make sure that the text is actually written to a file.

This can basically be done by calling reset towards the beginning of
mon_refresh_display and then calling finish at the end, regardless of
format.  This does mean the first time reset gets called will be right
after out->init, but that doesn't do anything bad.

Note that I am still setting output format to none for daemonize+curses.
This combination does not make sense.