Refactor: libcib, tools: Use the formatted output password prompt.
This replaces the password prompt that was built into cib_remote.c.
Calling it requires having an output object set in the CIB.  However, we
don't need this for anything except the remote case and we don't want
anyone else to use it, so it's hidden in the opaque data.  That then
means we need a private API function to do the setting.