Build: scheduler: allow 2.0 compatibility mode for certain output
Create a new --enable-compat-2.0 configure option to revert certain output
changes in 2.1.0, to support versions of tools and resource agents that rely on
parsing text output as it was in the 2.0 release series.

When building RPMs, define the new option on Fedora 34 and earlier and RHEL 8
and earlier, as pcs and resource-agents on those platforms are expected to
require it.

As of this commit, this is limited to whether agents are displayed as
ocf::PROVIDER:AGENT or ocf:PROVIDER:AGENT, but more differences are expected.