Log: libcrmcommon: be consistent with IPC server names in messages
Always refer to it as "<name> IPC".

A few messages incorrectly used the IPC server name as the client name, so fix
those. Rename the struct member from ->name to ->server_name to prevent future
mistakes.

Update CTS appropriately, including ignoring a "provider disconnected" message
that can be logged instead of "Connection ... failed" when the client is
waiting on a reply.