Fix: libcrmcommon: Simplify reset function for xml and html.
Instead of basically doing what finish would do and then resetting the
internal structures, just do the reset.  Leave it up to the caller to
call finish if they want it.  This makes these formatters act more like
text, which makes it easier to implement daemonize mode in crm_mon for
those formatters.