Low: Fix a problem with crm_resource exit code handling.
If no output is produced but an error message is printed (like, when an
inactive resource is provided on the command line), don't print the
error message for the pcmk_rc_no_output error code.  It's weird to see
output and a message about no output at the same time.

Similarly, don't print an "Error: OK" message when usage is printed.