Fix: fencer: update outdated pending operations according to returned ones from remote peer history
If a fencer had a pending fencing operation in the history but somehow
was not reachable when the fencing operation returned, the pending
operation would remain outdated and get stuck in the history.

The solution here is to broadcast returned fencing operations and update
outdated pending ones in the history.