Build: libcrmcommon,libcrmd: use version number consistently
Make sure we use the same Pacemaker version everywhere.

The configure script defines three separate variables for the version:

* VERSION and PACKAGE_VERSION: These are automatically generated by the
  AC_INIT() call, using the version number defined in m4/version.m4.
  AC_INIT() will automatically make these into substitution variables, so
  stop explicitly doing that. It also adds them to config.h (and there's no way
  to tell it not to).

* PACEMAKER_VERSION: This defaults to the above, but can be changed via the
  configure script's --version option. This is part of the public API due to
  being in include/crm_config.h.in.

We already redefine PACKAGE_VERSION to PACEMAKER_VERSION, which affects the
substitution variable as well as later uses in configure.ac, but unfortunately
it cannot affect the config.h constant, which means we should only use
PACEMAKER_VERSION in C.

We also now redefine VERSION to PACEMAKER_VERSION, to keep everything
consistent.

(BUILD_VERSION is a separate constant that usually has the git commit,
and is unchanged here.)