Test: scheduler: Clone placement when recovering promoted instance
Currently, when a promoted instance of an anonymous clone is stopped and
about to be recovered, the instances get shuffled. Instance 0, which is
running in non-promoted state on another node, gets moved to the
to-be-started-and-promoted node. Then instance 1 starts on the
non-promoted node. This causes an unnecessary restart on the
non-promoted node.

This will be fixed in an upcoming commit.

This commit adds six tests:
  - clone-anon-recover (correct)
  - clone-group-anon-recover (correct)
  - promotable-anon-recover-promoted (incorrect)
  - promotable-anon-recover-non-promoted (correct)
  - promotable-group-anon-recover-promoted (questionable)
    - correct in that it doesn't shuffle instances, but maybe incorrect
      in that it promotes a different node after one monitor failure on
      the promoted node
  - promotable-group-anon-recover-non-promoted (correct)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>