Build,Doc,Test: all: raise minimum Python dependency to 3.4
This is part of an effort to raise our minimum dependency versions to ones we
can actually test in CI, so we can ensure we support what we claim to.

The target for versions is what's available in major Linux distribution
releases from about 5 years ago, such as Debian 8 "jessie", Ubuntu 15 "xenial",
and RHEL 7.2.