Doc: books: switch latex engine to xelatex (for PDF format)
sphinx's default latex engine, pdflatex, errors out on the latex created by
sphinx. They're aware of the issue and may change the default eventually, but
there are trade-offs with the alternatives (xelatex or lualatex) as well.

For our purposes, pdflatex doesn't work (at least with sphinx 1.7.6), and
xelatex does, so it's an easy choice.

While we're at it, update all the dependencies needed for the PDF format.