Build: sysconfig: install sysconfig files via make install
Previously, make install did nothing with sysconfig files, and the RPM spec
file installed them. It makes more sense for make install to do it, since it
can respect --with-configdir and will work for users that don't build RPMs.

Also, move the sysconfig files to etc/sysconfig in the source tree.