Refactor: scheduler: Remove print_node.
This debugging function is the last user of fns->print that we care
about (there's one more in crm_simulate, but that'll be going away soon
enough).  There's other ways to print out a node now, and we don't want
to be using the older printing code so it can be deprecated.