Feature: Use formatted output for scheduler node actions.
Note that the stage7 function sets up its own output object.  This
function only ever logs (instead of printing to the screen), but it can
be called from various places that may want to print to the screen as
well.  stage7 can't know where it's being called from, so it just makes
sense to construct a log-specific output object and use it.