Feature: xml: Move node-related XML schema into its own file.
This allows it to be shared between the crm_mon and future crm_simulate
schemas.  Note that only the most complicated node schema has been
broken out.  There's still the simpler element-node-with-attribute, but
I don't think that will be used anywhere else.

Also note that any schemas including this one will need to define their
own node list that has the reference.  The idea here is that somewhere
might want a schema for just a single node instance in the future.