Fix: libpe_status: Don't use a default node list in pe__show_node_weights.
If a non-NULL resource is given, don't override the passed in list of
nodes with rsc->allowed_nodes.  We are passing that list almost
everywhere already, so overriding is a rare surprise.

Additionally, if there are children nodes, we need to pass the list of
allowed_nodes for the children into recursive calls.  Keeping the
initial list changes various values in the output of cts/cts-scheduler.