Fix: scheduler: handle ticket last-granted dates after 2038
The ticket state has a field for an epoch timestamp representing the time the
ticket was last granted. Previously, this was parsed as a regular int,
resulting in a Y2038 issue. Now, it's parsed as a long long.

The cluster did not use last-granted itself other than a display value.