Low: CIB: constrain remote-tls-port and remote-clear-port to port range
Previously, any integer was allowed for remote-tls-port and remote-clear-port.
If the value was out of the valid range (0-65535), we'd try and fail to start
the server. Now, log a warning and don't bother trying.

A new internal API function, pcmk__scan_port(), is added to abstract this
check. It is also now used when parsing CIB_port, but we don't change behavior
for invalid values in that case -- they will fail when signon() is attempted.