Low: tools: treat invalid error codes as unknown in crm_error
Previously, crm_error used crm_atoi() (aka crm_parse_int()) to parse the
command-line arguments for error codes. This meant that if someone gave invalid
text, it would be mapped to error code -1 (e.g. "Operation not permitted" in
legacy codes).

Now, map it to INT_MIN, which is nearly certain to map to "unknown error".