Feature: API: PCMK_ALLOW_DEPRECATED enables control of API availability
By default, external code using Pacemaker headers is allowed to use
deprecated Pacemaker APIs. If PCMK_ALLOW_DEPRECATED is defined to 0 before
including any Pacemaker headers, deprecated APIs will be unavailable. It is
strongly recommended to leave it unset for production and release builds,
to avoid breakage when Pacemaker upgrades deprecate more APIs. It should be
defined to 0 only for development and testing builds, to check for usage of
currently deprecated APIs.