Low: controller: bound stonith-max-attempts to a minimum of 1
Previously, we check stonith-max-attempts for the literal string "INFINITY" and
otherwise used crm_parse_ll().

That did not handle invalid values well (it would use -1), did not check for
out-of-bounds values (i.e. less than 1), did not impose an upper bound of
INFINITY as is implied by allowing the string, and would not recognize
"+INFINITY" as a synonym for "INFINITY" (as is done everwhere else).

Now, it uses char2score(), which bounds to +/-INFINITY and accepts all the
customary literal strings, and then additionally imposes a lower bound of 1.
It continues to use 10 as the default, which will now also be used for invalid
values.