Fix: libpacemaker: Don't shuffle anonymous clone instances
Currently, anonymous clone instances may be shuffled under certain
conditions, causing an unnecessary resource downtime when an instance is
moved away from its current running node.

For example, this can happen when a stopped promotable instance is
scheduled to promote and the stickiness is lower than the promotion
score (see the promotable-anon-recover-promoted test). Instance 0 gets
allocated first and goes to the node that will be promoted, causing it
to relocate if it's already running somewhere else.

There are also some other corner cases that can trigger shuffling, like
the one in the clone-anon-no-shuffle-constraints test.

The fix is to allocate an instance to its current node during
pre-allocation if that node is going to receive an instance at all.

Previously, if instance:0 was running on node1 and got pre-allocated to
node2 due to node2 having a higher weight, we backed out and immediately
gave up on pre-allocating instance:0.

Now, if instance:0 is running on node1 and gets pre-allocated to node2,
we increment the "reserved" counter (to ensure we don't allocate the max
number of instances without node2 getting one), and we make node2
unavailable. If allocated + reserved < max, we try pre-allocating
instance:0 again with node2 out of the picture.

This commit also updates several tests that contain unnecessary instance
moves, and it updates scores files that changed due to the fix.

Resolves: RHBZ#1931023

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>